Trait machinae::State

pub trait State<A, E, F>: Sized {
    fn start(&mut self, _args: A) -> Result<Trans<Self>, E> { ... }

    fn resume(&mut self, _args: A) { ... }

    fn pause(&mut self, _args: A) { ... }

    fn stop(&mut self, _args: A) { ... }

    fn update(&mut self, _args: A) -> Result<Trans<Self>, E> { ... }

    fn fixed_update(&mut self, _args: A) -> Result<Trans<Self>, E> { ... }

    fn event(&mut self, _args: A, _event: F) -> Result<Trans<Self>, E> { ... }
}

State trait with several callbacks which allow state transitions. It's recommended that you use this with an enum; if you prefer trait objects, you should use DynState instead.

Provided Methods

fn start(&mut self, _args: A) -> Result<Trans<Self>, E>

This method will be called before any other method of this state. Also see StateMachine::start.

fn resume(&mut self, _args: A)

Resumes the State. This will be called only if the state has been paused before (e.g. after a state has been pushed and popped again).

fn pause(&mut self, _args: A)

A state will be paused after it pushed a new state. Also, pause is always called before stop.

fn stop(&mut self, _args: A)

A state gets stopped one moment before it gets dropped, either because it got popped, it did a switch or the state machine has shut down.

fn update(&mut self, _args: A) -> Result<Trans<Self>, E>

State updates are performed by StateMachine::update.

fn fixed_update(&mut self, _args: A) -> Result<Trans<Self>, E>

Fixed state updates are performed by StateMachine::fixed_update.

fn event(&mut self, _args: A, _event: F) -> Result<Trans<Self>, E>

This method allows supplying events to the state via StateMachine::event.